Add files via upload
This file is intended to decompress a complete bundle of tasks downloaded in zip format from Moodle in a single folder with a subfolder for each student. The task of each student (which is supposed to be in zip format but works for many of them, just in case) is decompressed in his/her subfolder as well.